Add components for forum breadcrumbs
