+++ packages/ui/src/forum/hooks/useThreadBreadcrumbs.ts
@@ -1,0 +1,16 @@
+import { useMemo } from 'react'
+
+import { useGetForumThreadBreadcrumbsQuery } from '../queries'
+import { asThreadBreadcrumbs } from '../types'
+
+export const useThreadBreadcrumbs = (id: string) => {
+  const { data, loading } = useGetForumThreadBreadcrumbsQuery({ variables: { where: { id } } })
+  const breadcrumbs = useMemo(
+    () => (data?.forumThreadByUniqueInput ? asThreadBreadcrumbs(data.forumThreadByUniqueInput) : undefined),
+    [data, loading]
+  )
+  return {
+    isLoading: loading,
+    ...breadcrumbs,
+  }
+}
